+++ breaditnow-auth-api/src/main/java/com/breaditnow/auth/global/exception/AuthErrorCode.java
@@ -11,18 +11,6 @@ public enum AuthErrorCode implements ErrorCode {
 	UNSUPPORTED_PROVIDER(HttpStatus.BAD_REQUEST, "BA001", "지원하지 않는 소셜입니다."),
 
 	/**
-	 * 토큰(BB000)
-	 */
-	OAUTH2_ACCESS_TOKEN_EXPIRED(HttpStatus.UNAUTHORIZED, "BB001", "소셜 Access Token이 만료되었습니다. 재로그인해주세요."),
-	TOKEN_INVALID(HttpStatus.UNAUTHORIZED, "BB002", "유효하지 않은 토큰입니다."),
-	TOKEN_EXPIRED(HttpStatus.UNAUTHORIZED, "BB003", "만료된 토큰입니다."),
-	TOKEN_UNSUPPORTED(HttpStatus.UNAUTHORIZED, "BB004", "지원되지 않는 토큰입니다."),
-	TOKEN_WRONG(HttpStatus.UNAUTHORIZED, "BB005", "잘못된 토큰 서명입니다."),
-	REFRESH_TOKEN_NOT_EXIST_IN_COOKIE(HttpStatus.UNAUTHORIZED, "BB006", "쿠키에 Refresh 토큰이 없습니다. 재로그인해주세요."),
-	REFRESH_TOKEN_EXPIRED(HttpStatus.UNAUTHORIZED, "BB007", "만료된 Refresh 토큰 입니다. 재로그인해주세요."),
-	TOKEN_NOT_MATCHED(HttpStatus.UNAUTHORIZED, "BB008", "토큰이 일치하지 않습니다."),
-
-	/**
 	 * 기타(BC000)
 	 */
 	SERIALIZATION_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "BC001", "AuthToken을 JSON으로 직렬화하는데 실패했습니다."),
