--- breaditnow-auth-api/src/main/java/com/breaditnow/auth/global/security/Role.java
@@ -1,6 +1,0 @@
-package com.breaditnow.auth.global.security;
-
-public enum Role {
-	CUSTOMER,
-	OWNER
-}
