+++ breaditnow-auth-api/src/main/java/com/breaditnow/auth/global/security/oauth2/handler/Oauth2AuthenticationSuccessHandler.java
@@ -1,7 +1,7 @@
 package com.breaditnow.auth.global.security.oauth2.handler;
 
-import static com.breaditnow.auth.domain.token.domain.AuthTokenType.*;
 import static com.breaditnow.auth.global.security.oauth2.cookie.CookieOAuth2AuthorizationRequestRepository.*;
+import static com.breaditnow.common.secutiry.jwt.AuthTokenType.*;
 
 import java.io.IOException;
 import java.util.Optional;
@@ -12,11 +12,11 @@ import org.springframework.security.web.authentication.SimpleUrlAuthenticationSu
 import org.springframework.stereotype.Component;
 import org.springframework.web.util.UriComponentsBuilder;
 
-import com.breaditnow.auth.domain.token.domain.AuthToken;
 import com.breaditnow.auth.domain.token.repository.AuthTokenRepository;
-import com.breaditnow.auth.global.security.AccountContext;
-import com.breaditnow.auth.global.security.jwt.JwtTokenCreator;
 import com.breaditnow.auth.global.security.oauth2.cookie.CookieOAuth2AuthorizationRequestRepository;
+import com.breaditnow.common.secutiry.AccountContext;
+import com.breaditnow.common.secutiry.jwt.AuthToken;
+import com.breaditnow.common.secutiry.jwt.JwtTokenCreator;
 import com.breaditnow.common.util.CookieUtil;
 import com.breaditnow.domain.customer.entity.Customer;
 import com.breaditnow.domain.customer.repository.CustomerRepository;
