+++ breaditnow-auth-api/src/main/java/com/breaditnow/auth/global/security/oauth2/service/CustomOAuth2UserService.java
@@ -7,8 +7,8 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 
 import com.breaditnow.auth.domain.customer.service.CustomerRegistrationService;
-import com.breaditnow.auth.global.security.AccountContext;
 import com.breaditnow.auth.global.security.oauth2.model.OAuth2UserInfo;
+import com.breaditnow.common.secutiry.AccountContext;
 import com.breaditnow.domain.customer.entity.Customer;
 
 import lombok.RequiredArgsConstructor;
