+++ breaditnow-common/build.gradle
@@ -1,6 +1,14 @@
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation "org.apache.commons:commons-lang3:3.12.0"
+
+    implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
+    implementation 'org.springframework.boot:spring-boot-starter-security'
+
+
+    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
+    implementation 'io.jsonwebtoken:jjwt-impl:0.11.5'
+    implementation 'io.jsonwebtoken:jjwt-jackson:0.11.5'
 }
 
 test {
