+++ breaditnow-common/src/main/java/com/breaditnow/common/secutiry/Role.java
@@ -1,0 +1,6 @@
+package com.breaditnow.common.secutiry;
+
+public enum Role {
+	CUSTOMER,
+	OWNER
+}
