+++ breaditnow-common/src/main/java/com/breaditnow/common/secutiry/jwt/AuthTokenType.java
@@ -1,4 +1,4 @@
-package com.breaditnow.auth.domain.token.domain;
+package com.breaditnow.common.secutiry.jwt;
 
 public enum AuthTokenType {
 	ACCESS,
